// src/app/modules/admin/g-admin/sale-page/sale-page.types.ts
export interface salePagePagination {
    length: number;
    size: number;
    page: number;
    lastPage: number;
    startIndex: number;
    endIndex: number;
    pagination: {
        page: number,
        pageSize: number,
        pageCount: number,
        total: number
    }
}
// ...
export interface AssetType {

    id: string;
    name: string,
    asset_sizes: [
        {
            id: number,
            name: string,
            size: string,
        }
    ]
}